driver/httpvideodriver: implement screenshot
Implement screenshots for HTTP video cameras. For this we factor out the
default_port function out of stream() and reuse it for screenshot().

The screenshot function takes the stream, waits for 75 buffers to ensure
we have hit an i-frame in the pipeline so we can encode a complete
picture and than encodes the picture into a jpeg file.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>