doc: add environment file for the distributed case in getting_started
In this commit, I tried to pick up concepts like "targets", "drivers" and
"environment files" from the local example again in the distributed
example. This should help the reader put the new concepts in the
distributed case into context of what was explained before. This will
hopefully make it easier to understand the distributed case in general.

Signed-off-by: Leonardo Mörlein <git@irrelefant.net>