remote/exporter: add poll tracking override
The poll() function tracks whether the local information has changed vs
the previous state and only takes the availability into account. Since
we now poll() the resource on creation, the local state is immediately
updated with the information on the resource. This also means that
instead of a "clear" by sending an unavailable resource, we don't send an
update to the coordinator in case the resource is still available.

Previously:
Create Resource -> unavailable
poll() -> update to coordinator

Now:
create resource, poll() -> availability stays the same and we skip the
information update to the coordinator.

Add an explicit override to the poll function to mark the state as dirty
for the first poll().

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>