util/agents/network_interface: bail out if NM did not find an interface
nm.get_device_by_iface() returns None if there is no device for an
interface. This can happen if NetworkManager is not running, it it's in
an undefined state or if it does not manage the interface at all.

Instead of trying to call methods on the returned None type, let's bail
out early with a proper exception.

Signed-off-by: Bastian Krause <bst@pengutronix.de>