contrib/sync-places: filter tags during creation
In older versions of python, when iterating over a dictionary, a copy
would be created allowing us to modify the dictionary while we're
iterating over it. Newer versions however do not make this copy and now
raise runtime errors when the keys change while we're iterating over
them.

Transform the key/values during creation of the 'tags' dictionary to
avoid having to delete keys afterwards.

Signed-off-by: Bryan Brattlof <bb@ti.com>