strategy/common: introduce raise_if_broken decorator
Introduce a decorator meant to decorate the strategy's transition() and
force() methods. It stores the original exception in the broken
attribute and raises a StrategyError from it. All subsequent calls to the
decorated method will lead to the same exception.

This prevents a strategy from subsequent use after an exception
happened, thus no cascading errors can happen.

Signed-off-by: Bastian Krause <bst@pengutronix.de>