client: Clarify error message when unlock is required
Adding/removing aliases and matches requires a place to be idle. The
error when message when not minding this, is misleading in these cases
where really the place needs to be unlocked before the specific
operation.

Mention that the specific operation needs the place be idle. Instruct
the user to unlock if it's acquired by them.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>