doc/configuration: add more info on TFTP server requirements
While trying to make TFTPProviderDriver work with tftpd-hpa on Debian
Bookworm, I hit various different issues and ultimately couldn't really
make things work properly. The closest I got was by removing --secure
from the server's option but I would then need U-Boot to provide the
full path to the files and not relative to the TFTP server's root
directory, which makes things not go nicely in general and especially
not with labgrid (it uses external property as prefix for the path, and
not internal).

I installed atftpd instead and things worked out nicely out-of-the-box
on Debian Bookworm, so let's mention that in the docs along with the
requirements for the labgrid setup to work with the TFTPProvider

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>