driver: usbvideodriver: add support for overlaying FPS
We encourage addition of multiple quality settings for USB Video
devices, but there is no easy way to get feedback how many frames per
second are actually processed on the receiver side, which could
influence which quality settings would be chosen for Labgrid.

Let's add a fps parameter to USBVideoDriver::stream() that enables an
overlay with FPS statistics if enabled.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>