ConsoleProtocol: Update Signatures to match ConsoleExpectMixin
The ConsoleExpectMixin has learned some new tricks over time.
But the ConsoleProtocol was not updated accordingly.

Every class implementing the ConsoleProtocol also uses the
ConsoleExpectMixin - so it's safe to update the signatures according to
the ConsoleExpectMixin.

For ConsoleProtocol.expect() I've used the type definition of the
underlying pexpect library.

I have not updated the return types, since it's not clear to me what the
ConsoleExpectMixin should return.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>