remote/coordinator: drop returns from finally clauses
Exceptions in finally clauses containing return statements are
swallowed [1]. That means any exception raised in the try block and a
subsequent KeyError in the finally block leads to the original exception
being swallowed.

Since Python 3.14 "the compiler emits a SyntaxWarning when a return,
break or continue appears in a finally block (see PEP 765)." [2].

Fix this by moving the code relying on the session into the try clause
making the return in the except clause obsolete.

[1] https://peps.python.org/pep-0765/#motivation
[2] https://docs.python.org/3.14/reference/compound_stmts.html#finally-clause

Signed-off-by: Bastian Krause <bst@pengutronix.de>