driver/qemudriver: Add support for disk images and writer arguments
Add set_image() and set_writer_args() methods to QEMUDriver so that
the firmware writer can pass a disk image and extra QEMU arguments.
The image is attached as a virtio drive and the writer arguments are
appended to the QEMU command line. This allows EFI applications to
use persistent EFI variables via a pflash drive provided by the
writer.

Signed-off-by: Simon Glass <sjg@chromium.org>