driver: add RKUSBMaskromDriver bootstrap driver
Add a RKUSBMaskromDriver bootstrap driver with an accompanying
rkusbmaskrom agent to support bootstrapping targets with Rockchip SoCs.

The rkusbmaskrom agent expects the target to be in MASKROM mode and send
images to BootROM using vendor specific 0x471 and 0x472 control
transfers, in 4 KiB chunks.

The first image is loaded to SRAM using 0x471 and is expected to
initialize DRAM and then return back to BootROM. A second image can then
be loaded to start of DRAM using 0x472.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>