add NetworkCANPort resource and export
Add an exporter part for the remotely accessible CAN. As discussed, the
cannelloni is used as the underlying tunnel.

According to the cannelloni documentation, it should be used only in
environments where packet loss is tolerable. There is no guarantee that
CAN frames will reach their destination at all and/or in the right
order.

Signed-off-by: Tomas Novotny <tomas@novotny.cz>