driver/rawnetworkinterfacedriver: make start_record() block until capture started
The expectation of RawNetworkInterfaceDriver.start_record() is that it
returns once the packet capturing is running, so later code can start
right away generating packets that are received on the interface.

This does not work reliably currently because the labgrid-raw-interface
helper (wrapping tcpdump) is started via subprocess.Popen(), but that
does not mean the capture is actually running just yet.

To fix this, wait until the subprocess emits "tcpdump: listening on",
which is tcpdump's way of telling that it's now actually capturing.

Signed-off-by: Bastian Krause <bst@pengutronix.de>