remote/client: make acquire/release operate on all RemotePlaces in env
If an environment config with multiple RemotePlaces is given via
`-c`/`--config`, only the first RemotePlace is used on acquire/lock and
release/unlock.

Having multiple RemotePlaces is meant for multi place testing, so all
RemotePlaces need to be acquired/released.

This is implemented in a best effort way: If an error occurs, the
remaining places are still tried to be acquired/released. A combined
error message is shown if the operation failed on multiple places.
When called with `-d`, tracebacks of all errors are shown. If any of the
operations failed, the process exits with 1.

CI pipelines should use this pattern:

```shell
export LG_ENV=env.yaml
labgrid-client unlock || true
labgrid-client lock
```

Signed-off-by: Bastian Krause <bst@pengutronix.de>