driver: rawnetworkinterfacedriver: Add managed attribute
Adds an attribute that controls if the driver fully manages the
network interface, which causes it to setup and teardown the interface
on activate and deactivate. This allows users to disable this behavior
if they are externally managing the interface.

Closes: #1728

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>
[bst: make documentation more consistent]
Signed-off-by: Bastian Krause <bst@pengutronix.de>
(cherry picked from commit fe4ec8b3728f5924d04b2a8db755d550ccea8879)