util/loop: add is_new_loop()
There is currently no way to find out whether ensure_event_loop() created
a new loop. This information is useful in scenarios where a resource is
created without a session in a sync context, because then it needs to
shut down the newly created loop in clean up code. But this must not
happen if a running loop is used.

Since ensure_event_loop() is already public API people use, add a
separate function checking whether the passed in loop is identical to
the loop stashed in the loop creation code in ensure_event_loop().

Signed-off-by: Bastian Krause <bst@pengutronix.de>