github/workflows/scheduled-unit-tests: allow running via workflow_dispatch
This allows us to manually trigger the workflow via the web interface,
the API or the GitHub cli tool:

  $ gh workflow run build --ref master

This does not benefit us too much right now, but it will once this change
makes it into stable branches. Then we can trigger runs for these stable
branches from a job on the master branch using e.g.:

  $ gh workflow run build --ref stable-25.0

This is required because scheduled runs are only executed for the main
branch of a repository.

This is how we run scheduled tests for the different yocto version
branches in meta-labgrid[1].

[1]: https://github.com/labgrid-project/meta-labgrid/

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>
(cherry picked from commit a2e08e92e35e32479c2d902e130ff30969a255a4)