add support for transparent net device access using network namespaces
Add support for network namespaces that can be used to give a client
access to a NetworkInterface on an exporter. The data is streamed over
SSH between a tap in the netns on the client and a macvtap on the
exporter.

An agent process is used to initialize the user and network namespaces
and to keep them open for further operations.

This is not optimised for high throughput or low latency, but for
flexibility and ease-of-use. For cases where performance is more
important, this could be extended to support a macvlan interface in a
network namespace on the exporter.

The approach is still a bit experimental, so we might need to change
things in incompatible ways.

Co-developed-by: Joshua Watt <Joshua.Watt@garmin.com>
Co-developed-by: Rouven Czerwinski <rouven.czerwinski@linaro.org>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>