remote/coordinator: prohibit multiple exporters with the same name
Since the switch to gRPC, we only check that we don't have a connection
from the same peer, but not with the same name, which would break
assumptions elsewhere in the code and confuse users.

Fix it by raising an exception in this case and shutting down both
sides of the ExporterStream.

Fixes: #1774

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>